Improve Error Message For Deluge RPC Adapter
This provides better error messages specifically when SSL configuration is wrong.

I also made a change to DeamonException:
Instead of having a list of error types, I changed it to accept a string resources.
This allows different adapters to report errors that might be unique to their implementation without poluting the DeamonException class with types that are not relevant to anyone else.